adding packaging formatter
sometimes, the version create by vcver will not
line up with the version that is resolved by pip.
This is due to the fact that pep440 dictates
that some versions are formatted to specific strings (such as beta to b0).

By using packaging's version formatting,
these issues are eliminated, and the version
specified can be used as a specific version
reference by pip or other systems were the
version uploaded must match the VERSION file
string exactly.