Alternate strategy for extracting git branch
The git branch was ot working appropriately in travis-ci. This is due to the
tendency to pull and detach the sha directly.

As a result, changing the logic to ensure that the branch is appropriately
extracted. In addition, fixing the unit test to handle if the tests are
executed on a non-master branch.